Split Events and their models to separate files
